requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
coloredlogs==15.0.1
Django==5.1.2
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.61.0
numpy==2.1.2
onnxruntime==1.21.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.0.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==5.29.4
Pygments==2.19.1
PyMatting==1.1.13
referencing==0.36.2
rembg==2.0.65
reportlab==4.3.1
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.1
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
